Move button/modal to SR Detail page
